refactor(coupon): 코드 리뷰 반영 - Kafka 이벤트 처리 안전성 개선
- CouponEventListener: domain 패키지에서 infrastructure/kafka/coupon 으로 이동
  (Kafka 인프라 의존은 infrastructure 레이어에 위치해야 함)
- CouponEventListener: 비즈니스 예외(UserCouponNotFound, AlreadyUsed)와
  인프라 예외를 분리하여 처리 — 비즈니스 예외는 경고 로그 후 스킵,
  인프라 예외는 RuntimeException 재발행하여 Kafka 재처리 유도
- PaymentService: Kafka 이벤트 발행을 직접 호출 방식에서 제거
- PaymentEventListener: @TransactionalEventListener(AFTER_COMMIT)으로
  PaymentCompletedEvent Kafka 발행 — 트랜잭션 커밋 후에만 이벤트 발행하여
  정합성 보장
